* Added simple API to start testing with * Removed redundant imports, and made readme clear * Fixed git ignore
